chore: rename mocks to literals for clarity
