+++ packages/core/types/src/modules/entity-service/params/attributes/literals.ts
